Django/requirements.txt
altgraph==0.17
asgiref==3.8.1
async-timeout==4.0.3
c-formatter-42==0.2.7
daphne==3.0.2
channels==3.0.5
channels-redis==4.2.0
Django==4.2.13
django-cors-headers==4.0.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
future==0.18.2
macholib==1.15.2
msgpack==1.0.8
norminette==3.3.55
pillow==10.3.0
psutil==5.9.8
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytz==2024.1
redis==5.0.5
rest-framework-simplejwt==0.0.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
uvicorn
requests